python/requirements.txt
numpy
opencv-python
pyserial
scikit-image
matplotlib